backend/requirements.txt
Django==2.2.16
django-environ==0.4.5
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
